Add modal to add new tasks
There is a good amount of stuff in this Modal popup. First of all, we
are defining this template as an Angular template using <script
type="text/ng-template">. The cool thing about Angular templates is they
can be loaded from anywhere: locally or remote. The URL of the template
is the unique identifier, and if the template is defined locally, it
will be fetched locally. Templates are a great way to separate layouts
and UIs, so we use them extensively.

We then set a header with a button to close the modal, and then set up
our content area. For the form, we are calling createTask(task) when the
form is submitted. The task that is passed to createTask is the object
corresponding to the entered form data. Since our text input has
ng-model="task.title", that text input will set the title property of
the task object.